+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Latest:
 
 ### v1.4.2.1
 **Key Updates**
- - Initial contract funding limit
+ - Initial contract funding limits
 
 **Bugs Fixed**
  - Unable to upload a new file if 'force' is set and no file exists to delete.